new validation schemas, util functions, multer cleanup func, userDeletion fix, user model - updated default values

+++ server/src/controllers/user/deleteUser.ts
@@ -3,7 +3,6 @@ import { Request, Response, NextFunction } from 'express';
 import { AppError } from '../../errors/AppError.ts';
 import User from '../../db/models/User.ts';
 
-
 export const deleteUser = async (
   req: Request,
   res: Response,
@@ -11,13 +10,20 @@ export const deleteUser = async (
 ) => {
   try {
     //at this point user is authenticated and authorized to perform deletion
-    const user = await User.findOneAndDelete({
-      _id: req.params.id
+    const userToDelete = await User.findOneAndDelete({
+      _id: req.params.id,
     }).exec();
 
-    if (!user) {
-      throw new AppError("User not found", 404)
+    if (!userToDelete) {
+      throw new AppError('User not found', 404);
     }
+    if (userToDelete.role === 2) {
+      throw new AppError('Cannot delete an admin user', 400);
+    }
+
+    const user = await User.findOneAndDelete({
+      _id: req.params.id,
+    }).exec();
 
     res.status(200).send();
   } catch (error: any) {
